requirements.txt
psutil
flask
pillow
mss
screeninfo
